refactor: global패키지 생성 및 auth 패키지 분리
