+++ src/main/java/roomescape/global/auth/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package roomescape.domain.login;
+package roomescape.global.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
