+++ src/main/java/roomescape/global/exception/ClientIllegalArgumentException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class ClientIllegalArgumentException extends RuntimeException {
